Abstract subscription state to component
